feat(link): live demo e correção para publicação
