+++ package.json
@@ -17,8 +17,9 @@
     "prettier": "prettier write .",
     "prettier:check": "prettier --check .",
     "start:dev": "concurrently npm:dev npm:start",
-    "publish:button": "npm publish dist/packages/button --access public",
-    "publish:bundle": "npm publish dist/packages/animalia-web-components --access public",
+    "publish:link": "npm publish ./dist/packages/link --access public",
+    "publish:button": "npm publish ./dist/packages/button --access public",
+    "publish:bundle": "npm publish ./dist/packages/animalia-web-components --access public",
     "publish:all": "npm run publish:button && npm run publish:bundle",
     "build:publish": "npm run build && npm run publish:all",
     "storybook": "start-storybook -p 6006",
