+++ packages/link/src/stories/code.mdx
@@ -1,4 +1,4 @@
-import LiveDemo from './test.tsx';
+import LiveDemo from './live-demo.tsx';
 
 ## Desenvolvimento
 
@@ -31,6 +31,7 @@ da seguinte forma:
 ```css
 ani-link {
   --text-color: green;
+  --text-color-visited: orange;
   --text-color-hover: black;
 }
 ```
