+++ packages/link/src/stories/link.stories.ts
@@ -72,11 +72,8 @@ export default {
     },
     openNewTab: {
       control: 'boolean',
-      description: `
-      Indica se o link deverá ser aberta em uma nova aba.\n
-      Sempre que utilizar essa propriedade, é importante informar isso ao usuário através da label.\n
-      Por exemplo: Storybook (abrir em uma nova aba).
-      `,
+      description:
+        'Indica se o link deverá ser aberta em uma nova aba. Sempre que utilizar essa propriedade, é importante informar isso ao usuário através da label. Por exemplo: Storybook (abrir em uma nova aba).',
       defaultValue: 'false',
       table: {
         category: 'Propriedades',
